Make !game.currentRoom error message more accurate
When I didn't guard setting `this.currentRoom` in `move`, the error
message told me to make sure my `start` method was correct. While
"Sometimes the error message is misleading" is a good thing for our
trainees to learn, I'm not sure this is the place, as there's already a
lot of new things to grasp in this exercise.